+++ src/example/guide/getting-started.tsx
@@ -1,4 +1,4 @@
-import { Button } from "../../components/button";
+import { Button } from "../../lib/components/button";
 
 export default function Page() {
 	return (
